# src/sittingface_charts/openapi/services/auth.openapi.yaml
openapi: 3.1.0
info:
  title: auth API
  description: MonkeyBrain auth service API
  version: 1.0.0
paths:
...
  /api/v1/users/:
    get:
      tags:
      - Users
      summary: List Users
      operationId: list_users_api_v1_users__get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserEntryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Users
      summary: Create User
      operationId: create_user_api_v1_users__post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEntryCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEntryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/users/by-employee/{employee_id}:
    get:
      tags:
      - Users
      summary: Get User By Employee Id
      operationId: get_user_by_employee_id_api_v1_users_by_employee__employee_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: employee_id
        in: path
        required: true
        schema:
          type: string
          title: Employee Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEntryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/users/by-department/{department}:
    get:
      tags:
      - Users
      summary: Get Users By Department
      operationId: get_users_by_department_api_v1_users_by_department__department__get
      security:
      - HTTPBearer: []
      parameters:
      - name: department
        in: path
        required: true
        schema:
          type: string
          title: Department
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserEntryResponse'
                title: Response Get Users By Department Api V1 Users By Department  Department  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/users/by-team/{team}:
    get:
      tags:
      - Users
      summary: Get Users By Team
      operationId: get_users_by_team_api_v1_users_by_team__team__get
      security:
      - HTTPBearer: []
      parameters:
      - name: team
        in: path
        required: true
        schema:
          type: string
          title: Team
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserEntryResponse'
                title: Response Get Users By Team Api V1 Users By Team  Team  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/users/{user_id}:
    get:
      tags:
      - Users
      summary: Get User
      operationId: get_user_api_v1_users__user_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEntryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Users
      summary: Update User
      operationId: update_user_api_v1_users__user_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEntryUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEntryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Users
      summary: Delete User
      operationId: delete_user_api_v1_users__user_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/users/by-email/{email}:
    get:
      tags:
      - Users
      summary: Get User By Email
      operationId: get_user_by_email_api_v1_users_by_email__email__get
      security:
      - HTTPBearer: []
      parameters:
      - name: email
        in: path
        required: true
        schema:
          type: string
          title: Email
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEntryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/users/by-username/{username}:
    get:
      tags:
      - Users
      summary: Get User By Username
      operationId: get_user_by_username_api_v1_users_by_username__username__get
      security:
      - HTTPBearer: []
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
          title: Username
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEntryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/roles/{role_id}/permissions/{permission_id}:
    post:
      tags:
      - Roles
      summary: Add Permission To Role
      operationId: add_permission_to_role_api_v1_roles__role_id__permissions__permission_id__post
      security:
      - HTTPBearer: []
      parameters:
      - name: role_id
        in: path
        required: true
        schema:
          type: string
          title: Role Id
      - name: permission_id
        in: path
        required: true
        schema:
          type: string
          title: Permission Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Roles
      summary: Remove Permission From Role
      operationId: remove_permission_from_role_api_v1_roles__role_id__permissions__permission_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: role_id
        in: path
        required: true
        schema:
          type: string
          title: Role Id
      - name: permission_id
        in: path
        required: true
        schema:
          type: string
          title: Permission Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/permissions/:
    get:
      tags:
      - Permissions
      summary: List Permissions
      operationId: list_permissions_api_v1_permissions__get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPermissionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Permissions
      summary: Create Permission
      operationId: create_permission_api_v1_permissions__post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/permissions/by-resource/{resource}:
    get:
      tags:
      - Permissions
      summary: List Permissions By Resource
      operationId: list_permissions_by_resource_api_v1_permissions_by_resource__resource__get
      parameters:
      - name: resource
        in: path
        required: true
        schema:
          type: string
          title: Resource
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PermissionResponse'
                title: Response List Permissions By Resource Api V1 Permissions By
                  Resource  Resource  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/permissions/{permission_id}:
    get:
      tags:
      - Permissions
      summary: Get Permission
      operationId: get_permission_api_v1_permissions__permission_id__get
      parameters:
      - name: permission_id
        in: path
        required: true
        schema:
          type: string
          title: Permission Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Permissions
      summary: Update Permission
      operationId: update_permission_api_v1_permissions__permission_id__patch
      parameters:
      - name: permission_id
        in: path
        required: true
        schema:
          type: string
          title: Permission Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Permissions
      summary: Delete Permission
      operationId: delete_permission_api_v1_permissions__permission_id__delete
      parameters:
      - name: permission_id
        in: path
        required: true
        schema:
          type: string
          title: Permission Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PaginatedPermissionResponse:
      properties:
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
        results:
          items:
            $ref: '#/components/schemas/PermissionResponse'
          type: array
          title: Results
      type: object
      required:
      - total
      - page
      - page_size
      - results
      title: PaginatedPermissionResponse
...
    PaginatedUserEntryResponse:
      properties:
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
        results:
          items:
            $ref: '#/components/schemas/UserEntryResponse'
          type: array
          title: Results
      type: object
      required:
      - total
      - page
      - page_size
      - results
      title: PaginatedUserEntryResponse
    PermissionCreate:
      properties:
        permission_id:
          type: string
          minLength: 1
          title: Permission Id
        role_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Role Id
        name:
          type: string
          minLength: 1
          title: Name
        resource:
          type: string
          minLength: 1
          title: Resource
        action:
          type: string
          minLength: 1
          title: Action
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        is_active:
          type: boolean
          title: Is Active
          default: true
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - permission_id
      - name
      - resource
      - action
      title: PermissionCreate
    PermissionResponse:
      properties:
        permission_id:
          type: string
          minLength: 1
          title: Permission Id
        role_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Role Id
        name:
          type: string
          minLength: 1
          title: Name
        resource:
          type: string
          minLength: 1
          title: Resource
        action:
          type: string
          minLength: 1
          title: Action
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        is_active:
          type: boolean
          title: Is Active
          default: true
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - permission_id
      - name
      - resource
      - action
      title: PermissionResponse
    PermissionUpdate:
      properties:
        role_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Role Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        resource:
          anyOf:
          - type: string
          - type: 'null'
          title: Resource
        action:
          anyOf:
          - type: string
          - type: 'null'
          title: Action
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      title: PermissionUpdate
...
    RoleResponse:
      properties:
        role_id:
          type: string
          minLength: 1
          title: Role Id
        name:
          type: string
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        permissions:
          items:
            type: string
          type: array
          title: Permissions
        is_active:
          type: boolean
          title: Is Active
          default: true
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - role_id
      - name
      title: RoleResponse
...
    UserEntryCreate:
      properties:
        user_id:
          type: string
          minLength: 1
          title: User Id
        employee_id:
          type: string
          title: Employee Id
        name:
          type: string
          minLength: 1
          title: Name
        department:
          anyOf:
          - type: string
          - type: 'null'
          title: Department
        team:
          anyOf:
          - type: string
          - type: 'null'
          title: Team
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        is_active:
          type: boolean
          title: Is Active
          default: false
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        password:
          type: string
          minLength: 8
          format: password
          title: Password
          writeOnly: true
      type: object
      required:
      - user_id
      - name
      - password
      title: UserEntryCreate
    UserEntryResponse:
      properties:
        user_id:
          type: string
          minLength: 1
          title: User Id
        employee_id:
          type: string
          title: Employee Id
        name:
          type: string
          minLength: 1
          title: Name
        department:
          anyOf:
          - type: string
          - type: 'null'
          title: Department
        team:
          anyOf:
          - type: string
          - type: 'null'
          title: Team
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        is_active:
          type: boolean
          title: Is Active
          default: false
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - user_id
      - name
      title: UserEntryResponse
      description: "Password is intentionally absent \u2014 never serialized back\
        \ to the client."
    UserEntryUpdate:
      properties:
        employee_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Employee Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        department:
          anyOf:
          - type: string
          - type: 'null'
          title: Department
        team:
          anyOf:
          - type: string
          - type: 'null'
          title: Team
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
        password:
          anyOf:
          - type: string
            minLength: 8
            format: password
            writeOnly: true
          - type: 'null'
          title: Password
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      title: UserEntryUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer